Osa 4 done

+++ blogilista/controllers/blogs.js
@@ -1,29 +1,49 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-blogsRouter.get('/', async (request, response) => {
+blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('author')
-  response.json(blogs.map(b => b.toJSON()))
+  res.json(blogs.map(b => b.toJSON()))
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-  const user = await User.findById(request.body.userId)
-
-  const blog = new Blog({ ...request.body, author: user._id })
+blogsRouter.post('/', async (req, res, next) => {
+  const token = req.token
 
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({ ...req.body, author: user._id })
+
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
-    response.status(201).json(savedBlog)
+    res.status(201).json(savedBlog)
   } catch(exception) {
     next(exception)
   }
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+  const token = req.token
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blogToBeRemoved = await Blog.findById(req.params.id)
+
+    if(user._id.toString() !== blogToBeRemoved.author.toString()) {
+      return res.status(401).json({ error: 'User unauthorized to delete this item.' })
+    }
+
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
   } catch(e) {
